LayoutTests/fast/text-extraction/debug-text-extraction-role-suppression-expected.txt

root
    'Presentation role'
    'None role'
    'Generic role'
    label='Test group' 'Grouped content'
    'Rowgroup role'
    'Directory role'
    article 'Article content'
    navigation label='Site navigation'
        link uid=… 'Home'
    button uid=… 'Native button'
    role=banner 'Banner content'
    role=main 'Main content'
    uid=… role=button 'Custom button'
    role=status 'Status message'
    role=img label='Decorative image' '*'
    'Sidebar content'
    'Footer content'
    contentEditable uid=… role=textbox 'Editable content'
    role=dialog label='Test dialog' 'Dialog content'
    list role=tree
        'Tree item 1'
        'Tree item 2'
